broadband-modem-qmi: fix type comparison
Only applicable when building with WITH_NEWEST_QMI_COMMANDS.

  mm-broadband-modem-qmi.c: In function ‘process_common_info’:
  mm-broadband-modem-qmi.c:2924:20: error: comparison is always false due to limited range of data type [-Werror=type-limits]
   2924 |         if (mnc[2] == 0xFF) {
        |                    ^~

+++ src/mm-broadband-modem-qmi.c
@@ -2921,7 +2921,7 @@ process_common_info (QmiNasServiceStatus service_status,
     if (network_id_valid) {
         *mm_operator_id = g_malloc (7);
         memcpy (*mm_operator_id, mcc, 3);
-        if (mnc[2] == 0xFF) {
+        if ((guint8)mnc[2] == 0xFF) {
             memcpy (&((*mm_operator_id)[3]), mnc, 2);
             (*mm_operator_id)[5] = '\0';
         } else {
